Add files via upload
Slightly edit version of ryeon's commit. Takes out symbols (For some reason we were getting = as a keyword) and uses sklearn for calculating precision, f1, and recall. I took some words out of the index as I though the index is a little lengthy, but I don't think it did much. PoorIndex is the values I removed from index.